# docs/apispec_1.json
{
...
  "info": {
    "description": "A RESTful API for researchers to track arXiv papers, manage research notes, and use spaced repetition learning.",
    "termsOfService": "",
    "title": "ScholarTrack API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/notes": {
      "get": {
        "parameters": [
          {
            "description": "Filter by arXiv paper ID",
            "example": "2301.07041",
            "in": "query",
            "name": "paper_id",
            "required": false,
            "type": "string"
          },
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Results per page (max 100)",
            "in": "query",
            "name": "per_page",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of notes"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List all notes for the authenticated user.",
        "tags": [
          "Notes"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "content": {
                  "example": "The paper demonstrates that multi-head attention can be pruned without significant performance loss.",
                  "type": "string"
                },
                "paper_id": {
                  "description": "arXiv ID to link the note to a paper",
                  "example": "2301.07041",
                  "type": "string"
                },
                "title": {
                  "example": "Key findings on attention mechanisms",
                  "type": "string"
                }
              },
              "required": [
                "title",
                "content"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Note created successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Linked paper not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new research note.",
        "tags": [
          "Notes"
        ]
      }
    },
    "/api/notes/{note_id}": {
      "delete": {
        "parameters": [
          {
            "description": "The note ID",
            "example": 1,
            "in": "path",
            "name": "note_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Note deleted"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete a note.",
        "tags": [
          "Notes"
        ]
      },
      "get": {
        "parameters": [
          {
            "description": "The note ID",
            "example": 1,
            "in": "path",
            "name": "note_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Note details"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get a specific note by ID.",
        "tags": [
          "Notes"
        ]
      },
      "put": {
        "parameters": [
          {
            "description": "The note ID",
            "example": 1,
            "in": "path",
            "name": "note_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "content": {
                  "example": "Revised findings after reading the supplementary material.",
                  "type": "string"
                },
                "title": {
                  "example": "Updated title on attention mechanisms",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note updated successfully"
          },
          "400": {
            "description": "Request body is required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Update an existing note.",
        "tags": [
          "Notes"
        ]
      }
    },
    "/api/notes/{note_id}/flashcards": {
      "get": {
        "parameters": [
          {
            "description": "Note ID",
            "in": "path",
            "name": "note_id",
            "required": true,
            "type": "integer"
          },
          {
            "default": 1,
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "default": 20,
            "in": "query",
            "name": "per_page",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of flashcards for the note"
          },
          "401": {
            "description": "Missing or invalid JWT token"
          },
          "403": {
            "description": "Note does not belong to current user"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List flashcards for a specific note.",
        "tags": [
          "Flashcards"
        ]
      }
    },
...
    "/api/papers/{arxiv_id}": {
      "get": {
        "parameters": [
          {
            "description": "The arXiv paper identifier",
            "example": "2301.07041",
            "in": "path",
            "name": "arxiv_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Paper details (includes is_saved field if authenticated)"
          },
          "404": {
            "description": "Paper not found"
          }
        },
        "summary": "Get details of a specific paper by arXiv ID.",
        "tags": [
          "Paper Discovery"
        ]
      }
    },
    "/api/papers/{arxiv_id}/notes": {
      "get": {
        "parameters": [
          {
            "description": "The arXiv paper identifier",
            "example": "2301.07041",
            "in": "path",
            "name": "arxiv_id",
            "required": true,
            "type": "string"
          },
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Results per page (max 100)",
            "in": "query",
            "name": "per_page",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of notes for the paper"
          },
          "404": {
            "description": "Paper not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List all notes for a specific paper.",
        "tags": [
          "Notes"
        ]
      }
    },
...
    "/api/search/notes": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "n_results": {
                  "default": 5,
                  "description": "Number of results to return (max 50)",
                  "type": "integer"
                },
                "query": {
                  "example": "data augmentation techniques",
                  "type": "string"
                }
              },
              "required": [
                "query"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Semantic search results from user notes",
            "schema": {
              "properties": {
                "data": {
                  "items": {
                    "properties": {
                      "content": {
                        "type": "string"
                      },
                      "id": {
                        "type": "integer"
                      },
                      "relevance_score": {
                        "type": "number"
                      },
                      "title": {
                        "type": "string"
                      },
                      "type": {
                        "example": "note",
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                },
                "query": {
                  "type": "string"
                },
                "source": {
                  "example": "user_notes",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Missing or empty query field"
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Semantic search over the current user's notes.",
        "tags": [
          "Semantic Search"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT Bearer token. Format: \"Bearer {token}\"",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}